+++ src/icemac/recurrence/weekly.py
@@ -1,5 +1,5 @@
 from .base import StaticIntervalBase, next_date_of_same_weekday
-from datetime import timedelta
+from datetime import timedelta, datetime
 from icemac.recurrence.i18n import _
 import grokcore.component as grok
 
@@ -38,6 +38,17 @@ class BiWeekly(SameWeekdayBase):
     weight = 11
     title = _('every other week, same weekday (e. g. each second Friday)')
     interval = TWO_WEEKS
+
+    def _get_start_date(self):
+        candiate = next_date_of_same_weekday(self.context, self.interval_start)
+        # To compute the interval the time and timezone must be equal:
+        floored_context = self.interval_start.tzinfo.localize(
+            datetime.combine(self.context, self.interval_start.time()))
+        if (candiate - floored_context).days % 14 != 0:
+            # odd number of weeks
+            candiate = next_date_of_same_weekday(
+                self.context, self.interval_start + ONE_WEEK)
+        return candiate
 
     @property
     def info(self):