+++ CHANGES.rst
@@ -2,10 +2,14 @@
  Changelog
 ===========
 
-1.4 (unreleased)
-================
+1.3.1 (unreleased)
+==================
 
-- Nothing changed yet.
+- Fix computation of biweekly recurrences: Previously the first recurrence with
+  a matching weekday in the interval was used as the first result for the
+  interval. This is only correct in half of the cases. Now the computation of
+  the first recurrence in the interval takes into account that it has to be an
+  even number of weeks after the recurrence start date.
 
 
 1.3 (2017-01-07)